auth-service/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-database-url==2.2.0
psycopg2-binary==2.9.9
drf-spectacular==0.27.2
